property map class & guild properties
